refactor(copilot): extract entitlements registry with add-an-entitlement recipe
