+++ apps/sim/lib/workflows/custom-blocks/operations.ts
@@ -7,10 +7,8 @@ import {
   workspace,
 } from '@sim/db/schema'
 import { createLogger } from '@sim/logger'
-import { getErrorMessage } from '@sim/utils/errors'
 import { generateId, generateShortId } from '@sim/utils/id'
 import { and, eq, isNull, sql } from 'drizzle-orm'
-import { LRUCache } from 'lru-cache'
 import { isOrganizationOnEnterprisePlan } from '@/lib/billing/core/subscription'
 import { isFeatureEnabled } from '@/lib/core/config/feature-flags'
 import { extractInputFieldsFromBlocks, type WorkflowInputField } from '@/lib/workflows/input-format'
@@ -46,44 +44,15 @@ async function eligibleOrgForWorkspace(
 }
 
 /**
- * Cross-repo contract: the mothership (Go) reads this exact string from the chat
- * payload's `entitlements` array and matches it against
- * `core.EntitlementCustomBlocks` to gate the deploy agent's custom-block tool,
- * prompt section, and skill. The entitlement is the materialized form of
- * "`deploy-as-block` flag AND enterprise plan" for the workspace's org.
+ * Whether the workspace's org may use custom blocks (`deploy-as-block` flag +
+ * enterprise plan). Feeds the 'custom-blocks' entitlement in
+ * `@/lib/copilot/entitlements` and matches the REST route gates.
  */
-export const CUSTOM_BLOCKS_ENTITLEMENT = 'custom-blocks'
-
-const entitlementsCache = new LRUCache<string, Promise<string[]>>({
-  max: 500,
-  ttl: 5_000,
-})
-
-/**
- * The plan/flag-gated capabilities to send to the mothership for a request in
- * this workspace. Cached briefly so the several per-message callers (payload
- * build, headless execute, inbox) collapse to one evaluation.
- */
-export function computeWorkspaceEntitlements(
+export async function isCustomBlocksEligible(
   workspaceId: string,
   userId?: string
-): Promise<string[]> {
-  const cacheKey = `${workspaceId}:${userId ?? ''}`
-  const cached = entitlementsCache.get(cacheKey)
-  if (cached) return cached
-
-  const promise = eligibleOrgForWorkspace(workspaceId, userId)
-    .then((organizationId) => (organizationId ? [CUSTOM_BLOCKS_ENTITLEMENT] : []))
-    .catch((error) => {
-      entitlementsCache.delete(cacheKey)
-      logger.warn('Failed to compute workspace entitlements', {
-        workspaceId,
-        error: getErrorMessage(error),
-      })
-      return []
-    })
-  entitlementsCache.set(cacheKey, promise)
-  return promise
+): Promise<boolean> {
+  return (await eligibleOrgForWorkspace(workspaceId, userId)) !== null
 }
 
 /** A persisted custom block plus its live-derived Start input fields. */
